+++ Src/Private/Get-DiagBackupToRepo.ps1
@@ -5,7 +5,7 @@ function Get-DiagBackupToRepo {
     .DESCRIPTION
         Build a diagram of the configuration of Veeam VBR in PDF/PNG/SVG formats using Psgraph.
     .NOTES
-        Version:        0.1.0
+        Version:        0.4.0
         Author:         Jonathan Colon
         Twitter:        @jcolonfzenpr
         Github:         rebelinux
@@ -40,15 +40,15 @@ function Get-DiagBackupToRepo {
                                 node LocalReposDummy @{Label='Local Repository'; fontsize=18; fontname="Comic Sans MS bold"; fontcolor='#005f4b'; shape='plain'}
                                 $Rank = @()
                                 if ($LocalBackupRepo.count -le 4) {
-                                    foreach ($REPOOBJ in ($LocalBackupRepo | Sort-Object)) {
+                                    foreach ($REPOOBJ in ($LocalBackupRepo | Sort-Object -Property Name)) {
                                         $REPOHASHTABLE = @{}
                                         $REPOOBJ.psobject.properties | ForEach-Object {$REPOHASHTABLE[$_.Name] = $_.Value }
                                         node $REPOOBJ -NodeScript {$_.Name} @{Label=$REPOHASHTABLE.Label}
                                     }
-                                    edge -from BackupRepository -to $LocalBackupRepo.Name @{minlen=1; style=$EdgeDebug.style; color=$EdgeDebug.color}
+                                    edge -from LocalReposDummy -to $LocalBackupRepo.Name @{minlen=1; style=$EdgeDebug.style; color=$EdgeDebug.color}
                                 }
                                 else {
-                                    $Group = Split-array -inArray $LocalBackupRepo -size 4
+                                    $Group = Split-array -inArray ($LocalBackupRepo | Sort-Object -Property Name) -size 4
                                     $Number = 0
                                     while ($Number -ne $Group.Length) {
                                         SubGraph "LocalBackupGroup$($Number)" -Attributes @{Label=' '; style=$SubGraphDebug.style; color=$SubGraphDebug.color; fontsize=18; penwidth=1} {
@@ -77,7 +77,7 @@ function Get-DiagBackupToRepo {
                             SubGraph RemoteRepos -Attributes @{Label=' '; fontsize=18; penwidth=1.5; labelloc='t'; style='dashed'} {
                                 node RemoteReposDummy @{Label='Deduplicating Storage Appliances'; fontsize=18; fontname="Comic Sans MS bold"; fontcolor='#005f4b'; shape='plain'}
                                 if ($RemoteBackupRepo.count -le 4) {
-                                    foreach ($REPOOBJ in ($RemoteBackupRepo | Sort-Object)) {
+                                    foreach ($REPOOBJ in ($RemoteBackupRepo | Sort-Object -Property Name)) {
                                         $REPOHASHTABLE = @{}
                                         $REPOOBJ.psobject.properties | ForEach-Object { $REPOHASHTABLE[$_.Name] = $_.Value }
                                         node $REPOOBJ -NodeScript {$_.Name} @{Label=$REPOHASHTABLE.Label;}
@@ -85,7 +85,7 @@ function Get-DiagBackupToRepo {
                                     edge -from RemoteReposDummy -to $RemoteBackupRepo.Name @{minlen=1; style=$EdgeDebug.style; color=$EdgeDebug.color}
                                 }
                                 else {
-                                    $Group = Split-array -inArray $RemoteBackupRepo -size 4
+                                    $Group = Split-array -inArray ($RemoteBackupRepo| Sort-Object -Property Name) -size 4
                                     $Number = 0
                                     while ($Number -ne $Group.Length) {
                                         SubGraph "RemoteBackupRepo$($Number)" -Attributes @{Label=' '; style=$SubGraphDebug.style; color=$SubGraphDebug.color; fontsize=18; penwidth=1} {
@@ -115,7 +115,7 @@ function Get-DiagBackupToRepo {
                             SubGraph ObjectStorage -Attributes @{Label=' '; fontsize=18; penwidth=1.5; labelloc='t'; style='dashed'} {
                                 node ObjectStorageDummy @{Label='Object Repository'; fontsize=18; fontname="Comic Sans MS bold"; fontcolor='#005f4b'; shape='plain'}
                                 if ($ObjStorage.count -le 4) {
-                                    foreach ($STORAGEOBJ in ($ObjStorage | Sort-Object)) {
+                                    foreach ($STORAGEOBJ in ($ObjStorage | Sort-Object -Property Name)) {
                                         $OBJHASHTABLE = @{}
                                         $STORAGEOBJ.psobject.properties | ForEach-Object { $OBJHASHTABLE[$_.Name] = $_.Value }
                                         node $STORAGEOBJ -NodeScript {$_.Name} @{Label=$OBJHASHTABLE.Label}
@@ -123,7 +123,7 @@ function Get-DiagBackupToRepo {
                                     edge -from ObjectStorageDummy -to $ObjStorage.Name @{minlen=1; style=$EdgeDebug.style; color=$EdgeDebug.color}
                                 }
                                 else {
-                                    $Group = Split-array -inArray $ObjStorage -size 4
+                                    $Group = Split-array -inArray ($ObjStorage| Sort-Object -Property Name) -size 4
                                     $Number = 0
                                     while ($Number -ne $Group.Length) {
                                         SubGraph "ObjectStorage$($Number)" -Attributes @{Label=' '; style=$SubGraphDebug.style; color=$SubGraphDebug.color; fontsize=18; penwidth=1} {
@@ -152,7 +152,7 @@ function Get-DiagBackupToRepo {
                             SubGraph ArchiveObjectStorage -Attributes @{Label=' '; fontsize=18; penwidth=1.5; labelloc='t'; style='dashed'} {
                                 node ArchiveObjectStorageDummy @{Label='Archive Object Repository'; fontsize=18; fontname="Comic Sans MS bold"; fontcolor='#005f4b'; shape='plain'}
                                 if ($ArchiveObjStorage.count -le 4) {
-                                    foreach ($STORAGEArchiveOBJ in ($ArchiveObjStorage | Sort-Object)) {
+                                    foreach ($STORAGEArchiveOBJ in ($ArchiveObjStorage | Sort-Object -Property Name)) {
                                         $ARCHOBJHASHTABLE = @{}
                                         $STORAGEArchiveOBJ.psobject.properties | ForEach-Object { $ARCHOBJHASHTABLE[$_.Name] = $_.Value }
                                         node $STORAGEArchiveOBJ -NodeScript {$_.Name} @{Label=$ARCHOBJHASHTABLE.Label}
@@ -160,7 +160,7 @@ function Get-DiagBackupToRepo {
                                     edge -from ArchiveObjectStorageDummy -to $ArchiveObjStorage.Name @{minlen=1; style=$EdgeDebug.style; color=$EdgeDebug.color}
                                 }
                                 else {
-                                    $Group = Split-array -inArray $ArchiveObjStorage -size 4
+                                    $Group = Split-array -inArray ($ArchiveObjStorage| Sort-Object -Property Name) -size 4
                                     $Number = 0
                                     while ($Number -ne $Group.Length) {
                                         SubGraph "ArchiveObjectStorage$($Number)" -Attributes @{Label=' '; style=$SubGraphDebug.style; color=$SubGraphDebug.color; fontsize=18; penwidth=1} {