+++ Src/Private/Get-HtmlLabel.ps1
@@ -1,12 +1,21 @@
 Function Get-HTMLLabel {
     param(
         [string]$Label,
-        [string]$Type
+        [string]$Type,
+        [string]$PortPos
     )
 
-    if ($images[$Type]) {
+    if ($Type -eq 'NoIcon') {
+        $ICON = 'NoIcon'
+    }
+    elseif ($images[$Type]) {
         $ICON = $images[$Type]
     } else {$ICON = "no_icon.png"}
 
-    return "<TABLE border='0' cellborder='0' cellspacing='20' cellpadding='10'><TR><TD ALIGN='center' colspan='1'><img src='$($ICON)'/></TD></TR><TR><TD ALIGN='center'>$Label</TD></TR></TABLE>"
+    if ($ICON -ne 'NoIcon') {
+        return "<TABLE border='0' cellborder='0' cellspacing='20' cellpadding='10'><TR><TD ALIGN='center' colspan='1'><img src='$($ICON)'/></TD></TR><TR><TD ALIGN='center'>$Label</TD></TR></TABLE>"
+    } else {
+        return "<TABLE border='0' cellborder='0' cellspacing='20' cellpadding='10'><TR><TD ALIGN='center'>$Label</TD><TD PORT='port_child1'>Box</TD>
+        </TR></TABLE>"
+    }
 }