+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-# :arrows_clockwise: Microsoft Intune As Built Report Changelog
+# :arrows_clockwise: Veeam.Diagrammer Changelog
+
+## [0.4.0] - 2022-10.31
+
+### Added
+
+- Added support for Tape Backup Infrastructure diagramming
 
 ## [0.3.0] - 2022-10.17